Refactor architecture and home page function
