+++ main.js
@@ -1,7 +1,5 @@
-import './src/css/normalize.css'
-import './src/css/global.css'
-import Header from './src/modules/Header.js'
-import Hero from './src/modules/Hero.js';
+import './src/styles/normalize.css';
+import './src/styles/global.css';
+import Home from './src/pages/Home';
 
-Header();
-Hero();
+Home();
