+++ src/pages/Home.js
@@ -1,0 +1,9 @@
+import Header from '../../src/templates/Header.js';
+import Hero from '../../src/templates/Hero.js';
+
+const Home = () => {
+  Header();
+  Hero();
+};
+
+export default Home;
